Make library installable as a python pip package.
Right now this compiles only Kaldi by default. Kaldi is not right now
made a part of the pip package (that would be a bit odd...), but so
long as the install directory does not move, this solution works for
now for getting galvASR into a workflow depending on pip or conda.